Server: Implement PostgreSQL as a data provider for the server
We will be using PostgreSQL as our database engine for the server.

We use the configuration value 'POSTGRESQLCONNSTR_DefaultConnection' as
a connection string to the database.

We also have a temporary password 'temppw' to connect to the database.
In the future we will use a password file and direct users how to set up
the password.