Lib: Implement a MAUI library for performing grpc calls to the server
We want a library for easily performing our grpc calls to the server.

This allows us to have an easy to use and injectable services without
having to worry about the details of our api from any user of the
library.

We could also add versioning for the services if required.