Server: Fix issue with database seeding violating PK constraints
Apparently, by default you cannot have an ID of 0. However, we were
generating an ID of zero for a product which would cause bugs with psql
being able to create the database correctly.

We now use f.IndexVariable and pre-increment it as IndexVariable starts
at 0.

This now allows the seeding to be performed without any issues.