feat: Use TestContainers inside of our IntegrationTests for a real DB
We now use the TestContainers library to create a real database for our
integration tests.

We also have to seperate our github actions workflow to use linux vms
for our integration tests and windows vms for the other workflows. We
need windows vms for MAUI, however, we need linux for TestContainers to
be able to get the postgres container.